+++ R/DST_Runner.R
@@ -25,13 +25,18 @@
 #' A data stream pipe line
 #' consisting of a data stream, filters and a data mining task:
 #'
-#' ```DSD %>% DSF %>% DST```
+#' ```DSD %>% DSF %>% DST_Runner```
 #'
 #' Once the pipeline is defined, it can be run using [update()] where points are
-#' taken from the `DSD`, filtered through a sequence of `DSFs` and then used to update
-#' the task `DST`.
+#' taken from the [DSD] data stream source,
+#' filtered through a sequence of [DSF] filters and then used to update
+#' the [DST] task.
+#'
+#' [DST_Multi] can be used to update multiple models in the pipeline with the same stream.
 #'
 #' @family DST
+#' @name stream_pipeline
+#' @aliases stream_pipeline
 #'
 #' @param dsd A data stream (subclass of [DSD]) typically provided using a `%>%` (pipe).
 #' @param dst A data stream mining task (subclass of [DST]).
@@ -42,7 +47,7 @@
 #' # Set up a pipeline with a DSD data source, DSF Filters and then a DST task
 #' cluster_pipeline <- DSD_Gaussians(k = 3, d = 2) %>%
 #'                     DSF_Scale() %>%
-#'                     DST_Runner(DSC_DBSTREAM(r = .05))
+#'                     DST_Runner(DSC_DBSTREAM(r = .3))
 #'
 #' cluster_pipeline
 #'
@@ -50,11 +55,11 @@
 #' cluster_pipeline$dsd
 #' cluster_pipeline$dst
 #'
-#' # update the DST using the pipeline
+#' # update the DST using the pipeline, by default update returns the micro clusters
 #' update(cluster_pipeline, n = 1000)
 #'
 #' cluster_pipeline$dst
-#' get_centers(cluster_pipeline$dst)
+#' get_centers(cluster_pipeline$dst, type = "macro")
 #' plot(cluster_pipeline$dst)
 #' @export
 DST_Runner <- function(dsd, dst) {
@@ -71,8 +76,14 @@ DST_Runner <- function(dsd, dst) {
 }
 
 #' @export
-update.DST_Runner <- function(object, dsd = NULL, n = 1L, ...) {
-  if (!is.null(dsd))
-    stop("A dsd cannot be specified for update on DST_Runner!")
-  update(object$dst, object$dsd, n = n, ...)
+update.DST_Runner <- function(object, dsd = NULL, n = 1L, return = "model", ...) {
+
+
+  if (is.null(dsd))
+    ps <- get_points(object$dsd, n = n)
+  else
+    ps <- update(object$dsd, dsd, n = n)
+
+  #update(object$dst, object$dsd, n = n, return = return, ...)
+  update(object$dst, ps, n = n, return = return, ...)
 }